Rajout des tests

+++ src/main/java/org/loic/api_rest_team/services/impl/TeamServiceImpl.java
@@ -61,7 +61,12 @@ public class TeamServiceImpl implements TeamService {
 	@Override
 	public void deleteTeam(Long id) {
 		
-		this.teamRepository.deleteById(id);
+		boolean isExistBeforeDelete = this.teamRepository.findById(id).isPresent();
+		
+		if(isExistBeforeDelete) {
+			this.teamRepository.deleteById(id);
+		}
+		
 	}
 
 	@Override
